[Refactor] Nearby에서 쓰인 extension들 폴더에 분류
- Nearby에서 쓰인 extension들을 따로 분리하여 Extensions 폴더에 정리하였습니다.
